# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api QuickMenu",
        "description": "Documentação da API do sistema de restaurante quickmenu.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/menus": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Get all menu",
                "description": "Get all menu",
                "operationId": "e1e5e59287840410f132fe091d1fbfe7",
                "responses": {
                    "200": {
                        "description": "Lista de menus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Cardapio Teste"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Cardapio Teste"
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": "Teste.jpg"
                                            },
                                            "restaurant_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "active": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "slug": {
                                                "type": "string",
                                                "example": "cardapio-Teste"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-09-09T03:58:10.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-09-09T03:58:10.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Menu"
                ],
                "summary": "Store menu",
                "description": "Store menu",
                "operationId": "95eb7940157cb1f157df64b40ec5ccb0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Menu 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Menu 1"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "default": true,
                                        "example": true
                                    },
                                    "restaurant_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu registrado com sucesso"
                    }
                }
            }
        },
        "/api/menus/{id}": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Get menu",
                "description": "Get menu",
                "operationId": "b6d832555451425d6528be660b6bba4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Restaurante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID do Menu",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de menus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Cardapio Teste"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Cardapio Teste"
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": "Teste.jpg"
                                            },
                                            "restaurant_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "active": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "slug": {
                                                "type": "string",
                                                "example": "cardapio-Teste"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-09-09T03:58:10.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-09-09T03:58:10.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Menu"
                ],
                "summary": "Update menu",
                "description": "Update menu",
                "operationId": "760bce89fb01239625d38f9c46ae92e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Menu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Menu 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Menu 1"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "restaurant_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu registrado com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Menu"
                ],
                "summary": "Delete menu ",
                "description": "Deletar menu",
                "operationId": "ffa4606f2c88409876feda88cd424cc8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID do Menu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu deletado com sucesso"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Menu"
                ],
                "summary": "Destroy menu ",
                "description": "Desativar menu",
                "operationId": "12fa98b501dc4c5d02d6baddfb4608f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID do Menu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu desativado com sucesso"
                    }
                }
            }
        },
...
    },
...
}